+++ source/DSCResources/MSFT_SChannelSettings/MSFT_SChannelSettings.psm1
@@ -3,7 +3,9 @@ $script:modulesFolderPath = Join-Path -Path $script:resourceModulePath -ChildPat
 $script:resourceHelperModulePath = Join-Path -Path $script:modulesFolderPath -ChildPath 'SChannelDsc.Util'
 Import-Module -Name (Join-Path -Path $script:resourceHelperModulePath -ChildPath 'SChannelDsc.Util.psm1')
 
-$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_SChannelSettings'
+Import-Module -Name (Join-Path -Path $script:modulesFolderPath  -ChildPath 'DscResource.Common')
+
+$script:localizedData = SChannelDsc.Util\Get-LocalizedData -ResourceName 'MSFT_SChannelSettings'
 
 function Get-TargetResource
 {
@@ -289,6 +291,11 @@ function Get-TargetResource
     if ($null -eq $winhttpProtocols)
     {
         $winhttpProtocols = @()
+    }
+    else
+    {
+        # Fix for issue #28.
+        $winhttpProtocols = @($winhttpProtocols)
     }
 
     # FIPS Algorithm Policy